Move PeerListener to correct package
